requirements.txt
torch
torch-geometric
pytest
pandas
numpy
hydra-core
scipy
scikit-learn
pyarrow
PyYAML
tqdm
matplotlib
xgboost
catboost
omegaconf
pvlib
requests
pytz
seaborn
solcast